Added option to sort dimensions
If you are cutting your parts out of any kind of panelized material (i.e. plywood), you want the height of the part usually be the thickness of your material.
If you select this option, the dimensions are sorted to accommodate this, no matter how your part is oriented in your model.
The smallest value becomes the height (thickness), the next larger the width and the largest the length.